xum1541: Make payload-less control commands IN messages
For some reason (might be a driver bug) OUT control messages with a
zero-length payload return an error when using the libusb-win32
driver. IN control messages without a payload OTOH work fine.

So convert the OUT control messages to IN messages.

Unfortunately linux doesn't seem to like IN messages with a payload
length of zero (I didn't investigate if that's caused by libusb or
the kernel), so pass a one-byte dummy buffer.