xum1541_init(): Ignore some errors from set_configuration()
set_configuration() fails with LIBUSB_ERROR_IO on libusb
version 1.0.23 when using the libusb-win32 driver.

If a driver doesn't allow setting the configuration from user space
it's its own responsibility to do that. So assume that this is
what's happening and carry on.

Depending on how libusb will address this issue (if at all), a new
version of libusb might end up returning LIBUSB_ERROR_INVALID_PARAM
(that's what the current version returns for other "I know it won't
work so I won't even try" cases) so ignore that error, too.

As I consider LIBUSB_ERROR_NOT_SUPPORTED a better suited error
to return in this scenario (because I consider the parameter valid,
it's just that setting the configuration isn't supported),
pro-actively ignore that error too, in case the libusb maintainers
come to the same conclusion.