xum1541_init(): only configure device if necessary
Setting the usb configuration is a constant source of problems
(see commit message for f5c3bdf) while most of the time it's also
unnecessary, because normally the OS configures the device right
after it was plugged in.

So to avoid this problematic code path as much as possible, first
check if the xum1541 device is already configured and only call
set_configuration() if not.

Unfortunately libusb0's API has no get_configuration() call, so
in this case send a control message to the device to find out
if it is already configured.