xum1541_init(): only configure device if necessary
Setting the usb configuration is a constant source of problems
(see commit message for f5c3bdf) while most of the time it's also
unnecessary, because normally the OS configures the device right
after it was plugged in.

So to avoid this problematic code path as much as possible, first
check if the xum1541 device is already configured and only call
set_configuration() if not.

Unfortunately libusb0's API has no get_configuration() call, so
in this case send a control message to the device to find out
if it is already configured.

This change has the added benefit of making the set_interface()
call introduced by commit f5c3bdf redundant. While that solved
the problem of data toggles sometimes not being reset on linux,
it turned out to be a problem on MacOS with libusb 1.0.23 where
the set_interface() call not only fails but also triggers a
device reset: https://github.com/OpenCBM/OpenCBM/issues/39